feat: add support for cross-theme tech cards
