+++ src/components/pages/doc/technology-navigation/wrapper/index.js
@@ -1,0 +1,3 @@
+import Wrapper from './wrapper';
+
+export default Wrapper;
